refactor: split add_properties out of maintain.rs and factor its constraint gates
`mutation/maintain.rs` reached 2609 lines against the 2500-line ceiling
(main sits at 2490), and the constraint enforcement this sprint added
pushed `add_nodes` from 282 lines / 39 branches to 345 / 45, past its
recorded exception.

Two separate cuts, in the file that needed both:

- `add_properties` and its private vocabulary — the ancestor walk, the
  aggregate/spatial expression predicates, geometry resolution — move to
  `mutation/add_properties.rs`, the seam the existing `// ── add_properties ──`
  banner already marked. Nothing there writes topology, and no other
  operation in the file calls into it. `maintain.rs` drops to ~2040.

  `AddPropertiesReport` deliberately stays in `maintain.rs`: it is the
  return type of the public `kglite::api::mutation::add_properties`, so the
  pinned Rust API baseline records it at that path and moving it would
  change the public surface for a purely internal reshuffle.

- the per-row constraint work leaves `add_nodes`: `ConstraintColumns`
  gains `for_batch` (returns `None` for a type declaring nothing, which is
  what keeps the common bulk path free) and `gate_row`; `RowBuilder`
  captures the loop-invariant context so the row loop reads as "gate the
  row, build its action, queue it"; `describe_skipped_rows` takes over the
  skip reporting. `add_nodes` lands at 275 lines / 32 branches — below its
  recorded exception, which is tightened rather than raised.
